refactor: domain 패키지로 VoucherType 옮김
