+++ src/main/java/com/programmers/voucher/domain/VoucherType.java
@@ -1,8 +1,5 @@
-package com.programmers.voucher.enums;
+package com.programmers.voucher.domain;
 
-import com.programmers.voucher.domain.FixedAmountVoucher;
-import com.programmers.voucher.domain.PercentDiscountVoucher;
-import com.programmers.voucher.domain.Voucher;
 import com.programmers.voucher.request.VoucherCreationRequest;
 
 import java.util.Arrays;
